pyproject.toml
[project]
name = "cnn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=2.2.5",
    "ipykernel>=7.2.0",
    "kagglehub>=0.4.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.9.0",
    "tensorflow>=2.20.0",
]
